requirements.txt
streamlit==1.32.2
faiss-cpu==1.7.4
sentence-transformers==2.6.1
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
requests==2.31.0
